+++ eras/alonzo/impl/testlib/Test/Cardano/Ledger/Alonzo/Imp/UtxowSpec/Valid.hs
@@ -19,6 +19,7 @@ import Cardano.Ledger.Plutus (
  )
 import Control.Monad ((<=<))
 import Lens.Micro ((&), (.~))
+import Test.Cardano.Ledger.Alonzo.Arbitrary ()
 import Test.Cardano.Ledger.Alonzo.ImpTest
 import Test.Cardano.Ledger.Imp.Common
 import Test.Cardano.Ledger.Plutus.Examples
@@ -37,6 +38,7 @@ spec = describe "Valid transactions" $ do
           alwaysSucceedsWithDatumHash = hashPlutusScript $ alwaysSucceedsWithDatum slang :: ScriptHash (EraCrypto era)
           alwaysSucceedsNoDatumHash = hashPlutusScript $ alwaysSucceedsNoDatum slang :: ScriptHash (EraCrypto era)
           alwaysFailsWithDatumHash = hashPlutusScript $ alwaysFailsWithDatum slang :: ScriptHash (EraCrypto era)
+          alwaysFailsNoDatumHash = hashPlutusScript $ alwaysFailsNoDatum slang :: ScriptHash (EraCrypto era)
 
         it "Validating SPEND script" $ do
           txIn <- produceScript alwaysSucceedsWithDatumHash
@@ -68,14 +70,24 @@ spec = describe "Valid transactions" $ do
                 & inputsTxBodyL .~ [txIn]
                 & certsTxBodyL .~ [txCert]
 
-  it "Validating WITHDRAWAL script" $ do
-    const $ pendingWith "not implemented yet"
-  it "Not validating WITHDRAWAL script" $ do
-    const $ pendingWith "not implemented yet"
-  it "Validating MINT script" $ do
-    const $ pendingWith "not implemented yet"
-  it "Not validating MINT script" $ do
-    const $ pendingWith "not implemented yet"
+        it "Validating WITHDRAWAL script" $ do
+          account <- registerStakeCredential @era $ ScriptHashObj alwaysSucceedsNoDatumHash
+          expectTxSuccess <=< submitTx $
+            mkBasicTx $
+              mkBasicTxBody & withdrawalsTxBodyL .~ Withdrawals [(account, mempty)]
+
+        it "Not validating WITHDRAWAL script" $ do
+          account <- registerStakeCredential @era $ ScriptHashObj alwaysFailsNoDatumHash
+          expectTxSuccess <=< submitPhase2Invalid $
+            mkBasicTx $
+              mkBasicTxBody & withdrawalsTxBodyL .~ Withdrawals [(account, mempty)]
+
+        it "Validating MINT script" $ do
+          expectTxSuccess <=< submitTx <=< mkTokenMintingTx $ alwaysSucceedsNoDatumHash
+
+        it "Not validating MINT script" $ do
+          expectTxSuccess <=< submitPhase2Invalid <=< mkTokenMintingTx $ alwaysFailsNoDatumHash
+
   it "Validating scripts everywhere" $ do
     const $ pendingWith "not implemented yet"
   it "Acceptable supplimentary datum" $ do