+++ eras/mary/impl/testlib/Test/Cardano/Ledger/Mary/ImpTest.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE OverloadedLists #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableInstances #-}
@@ -8,6 +9,7 @@
 module Test.Cardano.Ledger.Mary.ImpTest (
   MaryEraImp,
   module Test.Cardano.Ledger.Allegra.ImpTest,
+  mkTokenMintingTx,
 ) where
 
 import Cardano.Crypto.DSIGN (DSIGNAlgorithm (..), Ed25519DSIGN)
@@ -18,8 +20,10 @@ import Cardano.Ledger.Crypto (Crypto (..))
 import Cardano.Ledger.Mary (MaryEra)
 import Cardano.Ledger.Mary.Core
 import Cardano.Ledger.Mary.Value
+import Lens.Micro ((&), (.~))
 import Test.Cardano.Ledger.Allegra.ImpTest
-import Test.Cardano.Ledger.Common
+import Test.Cardano.Ledger.Imp.Common
+import Test.Cardano.Ledger.Mary.Arbitrary ()
 import Test.Cardano.Ledger.Mary.TreeDiff ()
 
 instance
@@ -52,3 +56,15 @@ instance
   , Signable (DSIGN c) (Hash (HASH c) EraIndependentTxBody)
   ) =>
   MaryEraImp (MaryEra c)
+
+mkTokenMintingTx :: MaryEraImp era => ScriptHash (EraCrypto era) -> ImpTestM era (Tx era)
+mkTokenMintingTx sh = do
+  name <- arbitrary
+  count <- choose (1, 10)
+  let policyId = PolicyID sh
+  let ma = multiAssetFromList [(policyId, name, count)]
+  addr <- freshKeyAddr_
+  pure $
+    mkBasicTx mkBasicTxBody
+      & bodyTxL . mintTxBodyL .~ ma
+      & bodyTxL . outputsTxBodyL .~ [mkBasicTxOut addr (MaryValue mempty ma)]